Make /usr/bin/mupprnt executable when using simple.makefile
The mupprnt script is packaged in the tar file with non-executable permission
bits, so when it's copied to /usr/bin by using the simple.makefile, it lacks
executable permissions.  This flags warnings in Fedora package linters.

Using the ./configure; make; make install flow appears to install this script
with the proper executable bits.